+++ src/Fable.Transforms/Fable2Babel.fs
@@ -3379,7 +3379,8 @@ but thanks to the optimisation done below we get
                 ?isAbstract = isAbstract,
                 ?superClass = superClass,
                 typeParameters = typeParameters,
-                implements = implements
+                implements = implements,
+                ?doc = info.JsDoc
             )
         | FunctionExpression(_, parameters, body, returnType, typeParameters, _) ->
             Declaration.functionDeclaration (
@@ -3427,6 +3428,7 @@ but thanks to the optimisation done below we get
         ctx
         (ent: Fable.Entity)
         entName
+        (doc: string option)
         (consArgs: Parameter[])
         (consArgsModifiers: AccessModifier[])
         (consBody: BlockStatement)
@@ -3511,15 +3513,25 @@ but thanks to the optimisation done below we get
                 ?implements = implements
             )
 
-        ModuleDecl(entName, isPublic = ent.IsPublic)
+        ModuleDecl(entName, isPublic = ent.IsPublic, ?doc = doc)
         |> declareModuleMember com ctx classExpr
 
-    let declareClass (com: IBabelCompiler) ctx ent entName consArgs consBody superClass classMembers =
+    let declareClass
+        (com: IBabelCompiler)
+        ctx
+        (ent: Fable.Entity)
+        entName
+        doc
+        consArgs
+        consBody
+        superClass
+        classMembers
+        =
         if com.IsTypeScript then
             FSharp2Fable.Util.getEntityGenArgs ent |> makeTypeParamDecl com ctx |> Some
         else
             None
-        |> declareClassWithParams com ctx ent entName consArgs [||] consBody superClass classMembers
+        |> declareClassWithParams com ctx ent entName doc consArgs [||] consBody superClass classMembers
 
     let declareTypeReflection (com: IBabelCompiler) ctx (ent: Fable.Entity) entName : ModuleDeclaration =
         let ta =
@@ -3549,6 +3561,7 @@ but thanks to the optimisation done below we get
         ctx
         (ent: Fable.Entity)
         entName
+        doc
         (consArgs: Parameter[])
         (consBody: BlockStatement)
         baseExpr
@@ -3556,7 +3569,7 @@ but thanks to the optimisation done below we get
         : ModuleDeclaration list
         =
         let typeDeclaration =
-            declareClass com ctx ent entName consArgs consBody baseExpr classMembers
+            declareClass com ctx ent entName doc consArgs consBody baseExpr classMembers
 
         if com.Options.NoReflection then
             [ typeDeclaration ]
@@ -3710,7 +3723,14 @@ but thanks to the optimisation done below we get
                 yield makeMethod "Symbol.iterator" [||] (enumerableThisToIterator com ctx) returnType None
         |]
 
-    let transformUnion (com: IBabelCompiler) ctx (ent: Fable.Entity) (entName: string) classMembers =
+    let transformUnion
+        (com: IBabelCompiler)
+        ctx
+        (ent: Fable.Entity)
+        (entName: string)
+        (doc: string option)
+        classMembers
+        =
         let isPublic = ent.IsPublic
         let tagArgName = "Tag"
         let tagArgTa = makeAliasTypeAnnotation com ctx tagArgName
@@ -3798,6 +3818,7 @@ but thanks to the optimisation done below we get
                 ctx
                 ent
                 entName
+                doc
                 args
                 consBody
                 baseExpr
@@ -3948,6 +3969,7 @@ but thanks to the optimisation done below we get
                     ctx
                     ent
                     union_cons.Name
+                    doc
                     consArgs
                     consArgsModifiers
                     consBody
@@ -3976,7 +3998,7 @@ but thanks to the optimisation done below we get
                     |]
 
             let classMembers = Array.append [| cases |] classMembers
-            declareType com ctx ent entName args body baseExpr classMembers
+            declareType com ctx ent entName doc args body baseExpr classMembers
 
     let transformClassWithCompilerGeneratedConstructor
         (com: IBabelCompiler)
@@ -4021,7 +4043,7 @@ but thanks to the optimisation done below we get
                 Parameter.parameter (fi.Name, ?typeAnnotation = makeFieldAnnotationIfTypeScript com ctx fi.Type)
             )
 
-        declareType com ctx ent decl.Name args body baseExpr classMembers
+        declareType com ctx ent decl.Name decl.XmlDoc args body baseExpr classMembers
 
     let transformPojoDefinedByConsArgsToInterface
         (com: IBabelCompiler)
@@ -4319,7 +4341,7 @@ but thanks to the optimisation done below we get
             |> Option.defaultValue (None, consBody)
 
         [
-            yield! declareType com ctx ent classDecl.Name consArgs consBody baseExpr classMembers
+            yield! declareType com ctx ent classDecl.Name classDecl.XmlDoc consArgs consBody baseExpr classMembers
 
             if not ent.IsAbstractClass then
                 yield
@@ -4519,7 +4541,7 @@ but thanks to the optimisation done below we get
                         <| fun ctx -> transformClassWithPrimaryConstructor com ctx ent decl classMembers cons
                     | None ->
                         if ent.IsFSharpUnion then
-                            transformUnion com ctx ent decl.Name classMembers
+                            transformUnion com ctx ent decl.Name decl.XmlDoc classMembers
                         else
                             transformClassWithCompilerGeneratedConstructor com ctx ent decl classMembers